reorganize  state js files
